Refactor: split each tab into separate file
